fix: remove scrollbar-thumb disappearance and jitter on IE11
Removed the transition-duration property of the ScrollbarLayout/face CSS rule, to prevent jittering of the scrollbar thumb on IE11 while dragging it with the mouse.
Changed translateDOMPositionXY() to no longer hide the backface, in order to prevent the scrollbar thumb from disappearing on IE11 when dragging it with the mouse.