Fix scrollbarSpacer being pushed down.
See issue #291. When no scrolling is applied, ColumnsLeftShadow is rendered without any style but implicitly having height of the row.
Now, scrollbarSpacer is absolutely positioned and will be rendered below the ColumnsLeftShadow (normal div behavior)
The fix is to either make ColumnsLeftShadow be positioned absolutely so that it won't affect scrollbarSpacer, or just set top as 0.