Remove usage of componentWillReceiveProps (#612)
This removes the last usage of `componentWillReceiveProps` lifecycle
method from `FixedDataTableContainer`, by replacing it with `getDerivedStateFromProps`.

`getDerivedStateFromProps` dispatches a prop change action which updates the store based
on latest props. The new state of the store is then set on the component.

Note that `getDerivedStateFromProps` is called before every render, even for ones that are
caused by internal state updates. This means there's going to be another prop change action
being dispatched when the new store state is set, leading to cyclic updates.
We workaround this by maintaing a prop revision counter to ignore the state update caused by
the props change action, which is fine because `getDerivedStateFromProps` by definition
returns the new state.
object.