Introduce Rollup for packaging our distributables (#638)
The entire `internal` directory, and build scripts `buildNPMInternals.sh` and `buildDist.sh`
are removed.
We instead have rollup config files, named `rollup.config.js` and `cssRollup.config.js`
which is used by rollup to create the distributables.

For the CommonJS and ESM targets, the builds are present in dist/cjs and dist/es respectively.
The CSS and UMD builds are put in dist/ as before, for backward compatibility.

Now, there's also two entry points in package.json, one for CJS and one for ESM.

Other notes:
We'll no longer check out the dist folder to the repo.