Allow ReorderCell to work in React v19 (#745)
The `<ReorderCell />` plugin component doesn't work correctly in React v19.
This is because it uses the `ReactDOM.render` API which is removed in React v19.

## Fix
I'm adding in a new prop named `__react19RootCreator` which allows users to pass in the `createRoot` API so that `<ReorderCell />` works as expected.
The prop is already marked as deprecated in the docs, and prefixed with`__` to make it apparent that this is more of a hack/temporary workaround.
We'll remove it in FDT v3 since we'll probably only support React 18 and above.

Example usage:
```tsx
import { createRoot } from 'react-dom/client';

...

const reorderCell = (
  <Plugins.ReorderCell
    __react19RootCreator={createRoot}
  >
    {cellContents}
  </Plugins.ReorderCell> 
)
```